+++ include/ambition/Ambition_user_interface.hh
@@ -232,9 +232,32 @@ constexpr auto RALLY_POINT_BUTTON = Rectangle::fromPoint(
   }
 );
 
+const auto DISCORD_BUTTON = BOUNDS
+  .inner(4)
+  .internal(
+    { .width = 132, .height = 20 },
+    HorizontalAlignment::Left,
+    VerticalAlignment::Bottom
+  );
+const auto DISCORD_BUTTON_CLICK_AREA = DISCORD_BUTTON.outer(4, 2, 4, 4);
+
+const auto FEEDBACK_LINK = BOUNDS
+  .inner(0, 0, 0, DISCORD_BUTTON.outer(4, 2, 4, 4).height())
+  .inner(12, 0, 0, 2)
+  .internal(
+    { .width = 132, .height = 20 },
+    HorizontalAlignment::Left,
+    VerticalAlignment::Bottom
+  );
+const auto FEEDBACK_LINK_CLICK_AREA = FEEDBACK_LINK.outer(4, 4, 4, 2);
+
 
 Size bitmapSize(
   const char* bitmap
+);
+
+bool detectMouseClick(
+  const Rectangle area
 );
 
 void drawPanel(